Recycler view con BOTONES DENTRO, con ViewModel AVANZADO, FUNCIONES BASICAS EN LA BASE DE DATOS LISTADAPTER INTELIGENTE SNACKBAR EVENTOS

+++ app/src/main/java/com/example/recicle/MainActivityViewModel.kt
@@ -1,15 +1,46 @@
 package com.example.recicle
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.ViewModel
+import android.app.Application
+import android.view.View
+import androidx.lifecycle.*
 import data.Repository
 import data.entity.Student
+import util.Event
 
 //como le quiero pasar el repositorio tengo que crear un factory
-class MainActivityViewModel(val repository: Repository): ViewModel() {
+//variable applicacion para obtener recursos de la actividad
+class MainActivityViewModel(val repository: Repository,private val application: Application): ViewModel() {
+
+    val queryTrigger :MutableLiveData<Int> = MutableLiveData()
+    init {
+        queryTrigger.value=0
+    }
+  var students :LiveData<List<Student>> = queryTrigger.switchMap {
+      if(it == 0){
+          repository.queryAllStudents()
+      }else{
+          repository.queryMayores()
+      }
+  }
 
 
-    val students : LiveData<List<Student>> = repository.queryAllStudents()
+
+
+    //val students : LiveData<List<Student>> = Transformations.map(originalStudents){ originalList ->
+      //  originalList.filter { it.age<18 }
+            //.count()
+    //}
+
+
+    val lblEmtyViewVisibility: LiveData<Int> = Transformations.map(students){
+        if(it.isEmpty()) View.VISIBLE else View.INVISIBLE
+    }
+
+    //eventos
+    private val _onShowMessage : MutableLiveData<Event<String>> = MutableLiveData()//comunicacion entre la actividad y elviewmodel
+    val onShowMessage : LiveData<Event<String>>
+        get() = _onShowMessage
+
     fun addStudent(student: Student) {
         repository.addStudent(student)
 
@@ -21,8 +52,17 @@ class MainActivityViewModel(val repository: Repository): ViewModel() {
     }
 
     fun deleteStudent(student: Student) {
+        if(students.value?.isNotEmpty()==true){
+            repository.deleteStudent(student)
+        }else{
+            _onShowMessage.value =Event(application.getString(R.string.app_name))
+            //
+        }
 
-        repository.deleteStudent(student)
+    }
+
+    fun getMayores() {
+        queryTrigger.value=18
     }
